+++ ruma-client-api/CHANGELOG.md
@@ -24,7 +24,9 @@ Improvements:
 * Add method `into_event_content` for `r0::room::create_room::CreationContent`
 * Add room visibility endpoints: `r0::directory::{get_room_visibility, set_room_visibility}`.
 * Add is_empty helpers for structs in `r0::sync::sync_events`
-* Add a constructor for `r0::room::create_room::Request`
+* Add a constructor for request structs of the followign endpoints
+  * `r0::room::create_room`
+  * `r0::message::get_message_events`
 
 Deprecations: